fix: refactor md5 validate and test
