+++ pkg/util/md5/helper.go
